test: split MigrationUserTaskUpdateIT in separate classes
